@ fix create from timestamp with miliseconds

+++ src/DateTimeFromTimestamp.php
@@ -28,6 +28,6 @@ final class DateTimeFromTimestamp
      */
     public static function createIncludingMilliseconds(int $milliseconds): DateTimeImmutable
     {
-        return self::create($milliseconds / 1000);
+        return self::create((int)($milliseconds / 1000));
     }
 }
